Refactored nmc into two modules and added descriptive error names.
